PaaS/Development/Automation/requirements.txt
boto3>=1.26.0
azure-mgmt-compute>=29.0.0
google-cloud-compute>=1.10.0
kubernetes>=26.1.0
docker>=6.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
requests>=2.31.0
prometheus-client>=0.17.0
fastapi>=0.100.0
uvicorn>=0.23.0
psutil>=5.9.0
cryptography>=41.0.0
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
